Pequeño arreglo en registrarPaciente

+++ src/uiMain/Gestion/gestionPacientes/RegistrarPaciente.java
@@ -12,7 +12,7 @@ public class RegistrarPaciente {
         System.out.println("Ingrese el nombre del paciente:");
         String nombre = sc.next();
         System.out.println("Ingrese el número de cédula: ");
-        int id = sc.nextByte();
+        int id = sc.nextInt();
         System.out.println("Ingrese su tipo de eps: ");
         String eps = sc.next();
         Paciente paciente = new Paciente(id, nombre, eps);
